Now randomizes the output.
The string format is a hack, basically just make in string append
function.
However, the haskell format library I looked into, Text-format, has
terrible documentation & decided to name one type (Only) the same as
Sqlite-simple.
All of this made me decide to use the unsafe version because text-format
was too much of a pain to use.